% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_list_images_sorted.R
\name{.list_images_sorted}
\alias{.list_images_sorted}
\title{List image files from an folder sorted by their normalized name}
\usage{
.list_images_sorted(folder)
}
\arguments{
\item{folder}{Character string specifying a folder path. The folder should include image files.}
}
\value{
Date frame with the original file paths and the normalized names sorted by the normalized names
}
\description{
Lists all image files in a folder and returns them sorted by their
normalized file names.
}
