% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_normalize_name.R
\name{.normalize_name}
\alias{.normalize_name}
\title{Normalizing a file name}
\usage{
.normalize_name(path_or_name)
}
\arguments{
\item{path_or_name}{Character string containing a file name or a file path}
}
\value{
A normalized, lowercase file name without extensions
}
\description{
Extracting the base name from a file path,
Converting it to lower case and
Removing the file extension.
}
