% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_scales.R
\name{get_scales}
\alias{get_scales}
\title{Extracting the scale information from DM3 files belonging to each image}
\usage{
get_scales(folder)
}
\arguments{
\item{folder}{Path to the folder containing the image files and the corresponding DM3 files}
}
\value{
A named list. Each element corresponds to a image and contains:
\describe{
 \item{scale_x}{Pixel size in x direction}
 \item{scale_y}{Pixel size in y direction}
 \item{unit_x}{Unit of pixel size in x direction}
 \item{unit_y}{Unit of pixel size in y direction}
 \item{mm_per_pixel_x}{Pixel size in millimeters in x direction}
 \item{mm_per_pixel_y}{Pixel size in millimeters in y direction}
 \item{source}{Source of the scale information}
 }
}
\description{
Reads the pixel size and unit information from DM3 files associated with the
image files in a folder and returns the scale information converted to
millimeters per pixel
}
