% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{setup}
\alias{setup}
\title{Interactive environment setup}
\usage{
setup()
}
\value{
Invisibly returns \code{TRUE} if the setup is completed successfully.
}
\description{
Runs an interactive setup assistant to configure the Python / *Cellpose*
environment required for *NanopixR*. This function is intended for interactive
use only.
}
